pyproject.toml
[project]
name = "eo-flood-ops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "absl-py>=2.3.1",
    "geopandas>=1.1.1",
    "hydromt>=1.2.0",
    "ipykernel>=7.1.0",
    "matplotlib>=3.10.7",
    "opencv-python>=4.12.0.88",
    "pillow>=12.0.0",
    "pyinstaller>=6.16.0",
    "rasterio>=1.4.3",
    "ruff>=0.14.2",
    "scikit-image>=0.25.2",
    "scipy>=1.16.2",
    "xarray>=2025.10.1",
]
